Remove runtime usage of prisma
