+++ abods-api/src/types/extra.ts
@@ -1,6 +1,5 @@
 import express from "express";
 import { IncomingHttpHeaders } from "http";
-import { PrismaClient } from "@prisma/client";
 import { Kysely } from "kysely";
 import { DB } from "../kysely.js";
 import { BaseContext } from "@apollo/server";
@@ -14,7 +13,6 @@ export interface RequestContext extends BaseContext {
   req: express.Request;
   res: express.Response;
   headers: IncomingHttpHeaders;
-  db: PrismaClient;
   apiKeyAuth?: AuthContext;
   kysely: Kysely<DB>;
 }
